Move explicitly solvate to a species
